open_clip 실험환경/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
anyio==4.12.1
attrs==25.4.0
braceexpand==0.1.7
certifi==2026.2.25
charset-normalizer==3.4.4
click==8.3.1
clip-benchmark==1.6.2
cuda-bindings==12.9.4
cuda-pathfinder==1.2.2
cuda-toolkit==12.8.1
datasets==4.7.0
dill==0.4.0
dnspython==2.8.0
einops==0.8.2
filelock==3.20.0
frozenlist==1.8.0
fsspec==2025.12.0
ftfy==6.3.1
h11==0.16.0
hf-xet==1.3.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.5.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
markdown-it-py==4.0.0
MarkupSafe==3.0.2
-e git+https://github.com/microsoft/GLIP.git@9dda9558c1ef59bb6cdc8e896e2bcab775a68ff0#egg=maskrcnn_benchmark
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.18
networkx==3.6.1
ninja==1.13.0
nltk==3.9.3
numpy==2.3.5
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu11==9.1.0.70
nvidia-cudnn-cu12==9.19.0.56
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu11==2.21.5
nvidia-nccl-cu12==2.28.9
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.8.90
open_clip_torch==3.3.0
opencv-python==4.13.0.92
packaging==26.0
pandas==3.0.1
pillow==12.0.0
prettytable==3.17.0
propcache==0.4.1
protobuf==7.34.0
pyarrow==23.0.1
pycocoevalcap==1.2
pycocotools==2.0.11
Pygments==2.19.2
pymongo==4.16.0
python-dateutil==2.9.0.post0
PyYAML==6.0.3
regex==2026.2.28
requests==2.32.5
rich==14.3.3
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
setuptools==70.2.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
sympy==1.14.0
tensorboardX==2.6.4
threadpoolctl==3.6.0
timm==1.0.25
tokenizers==0.22.2
torch==2.12.0.dev20260305+cu128
torchaudio==2.11.0.dev20260305+cu128
torchvision==0.26.0.dev20260305+cu128
tqdm==4.67.3
transformers==5.3.0
triton==3.6.0+git9844da95
typer==0.24.1
typing_extensions==4.15.0
urllib3==2.6.3
wcwidth==0.6.0
webdataset==1.0.2
xxhash==3.6.0
yacs==0.1.8
yarl==1.23.0
